Add MPL-2.0 license
I like this license (except for the future-versions clause) and I don't
think it's incompatible with Rust's ecosystem of `MIT-or-APACHE-2.0`
crates.

Was considering using `MPL-2.0-no-copyleft-exception` but that would
cause problems for any GPL folks.

If this license causes problems for anyone I would be open to relicensing
to `MIT-or-APACHE-2.0` in the future.