Check results of scan in tests
Moves the `parent` and `child` tests from the `device` module to `scan`,
and updates all tests to use a local mock sysfs where we know what the
expected results will look like.

This mock sysfs contains a few devices, some of which are duplicates,
and some of which are parents and children of one another.